azurerequirement.txt
azure-core==1.11.0
azure-storage-blob==12.7.1
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
cryptography==3.1.1
idna==2.10
isodate==0.6.0
msrest==0.6.21
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.2
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
setuptools-rust==0.11.6
six==1.15.0
toml==0.10.2
urllib3==1.26.3
